refactor: plan/month 안의 내용을 plan으로 이동 (#41)
